pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Audio Splitter Web Application with Enhanced Security"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "ffmpeg-python>=0.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pydub>=0.25.1",
    "python-magic>=0.4.27",
]
